refactor: consolidate supabase client
