+++ src/components/NegotiationHistory.tsx
@@ -2,7 +2,7 @@ import { useState, useEffect } from 'react';
 import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';
 import { Badge } from '@/components/ui/badge';
 import { Button } from '@/components/ui/button';
-import { supabase } from '@/lib/supabase';
+import { supabase } from '@/lib/supabase-enhanced';
 import { MessageCircle, Clock, CheckCircle, XCircle } from 'lucide-react';
 
 interface NegotiationRecord {
