+++ src/components/funding/FundingHub.tsx
@@ -3,7 +3,7 @@ import { Button } from '@/components/ui/button';
 import { Input } from '@/components/ui/input';
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '@/components/ui/select';
 import { OpportunityCard } from './OpportunityCard';
-import { supabase } from '@/lib/supabase';
+import { supabase } from '@/lib/supabase-enhanced';
 import { Search, Filter, Sparkles } from 'lucide-react';
 import { InvestmentTips } from '@/components/InvestmentTips';
