+++ src/components/investor/CoInvestmentHub.tsx
@@ -6,7 +6,7 @@ import { Progress } from '@/components/ui/progress';
 import { Input } from '@/components/ui/input';
 import { Textarea } from '@/components/ui/textarea';
 import { Dialog, DialogContent, DialogHeader, DialogTitle, DialogTrigger } from '@/components/ui/dialog';
-import { supabase } from '@/lib/supabase';
+import { supabase } from '@/lib/supabase-enhanced';
 import { useAppContext } from '@/contexts/AppContext';
 import { Users, TrendingUp, Calendar, Plus, DollarSign } from 'lucide-react';
